requirements.txt
requests
gitpython
openai
tenacity
PyJWT
cryptography
